attempted fix for manual boost compilation against non-global python

+++ virtualenv-bootstrap.sh
@@ -471,10 +471,14 @@ if [ "$?" == 65 ]; then
     #boost not found
     echo "boost-python not found for this version of Python, we are gonna attempt to compile it manually"
     TS=`date +%s`
-    wget "http://downloads.sourceforge.net/project/boost/boost/1.58.0/boost_1_58_0.tar.bz2?r=http%3A%2F%2Fsourceforge.net%2Fprojects%2Fboost%2Ffiles%2Fboost%2F1.58.0%2F&ts=$TS&use_mirror=garr" -O boost.tar.bz2
-    tar -xjf boost.tar.bz2 2>/dev/null
+    if [ ! -f boost.tar.bz2 ]; then
+        wget "http://downloads.sourceforge.net/project/boost/boost/1.58.0/boost_1_58_0.tar.bz2?r=http%3A%2F%2Fsourceforge.net%2Fprojects%2Fboost%2Ffiles%2Fboost%2F1.58.0%2F&ts=$TS&use_mirror=garr" -O boost.tar.bz2
+        tar -xjf boost.tar.bz2 2>/dev/null
+    fi
     cd boost*
-    ./bootstrap.sh --with-libraries=python --prefix=$VIRTUAL_ENV --with-python-root=$VIRTUAL_ENV
+    PYTHONINCLUDE=`ls -ld $VIRTUAL_ENV/include/python3*`
+    export CPLUS_INCLUDE_PATH="$PYTHONINCLUDE:$CPLUS_INCLUDE_PATH"
+    ./bootstrap.sh --with-libraries=python --prefix=$VIRTUAL_ENV --with-python-root=$VIRTUAL_ENV --with-python="$VIRTUAL_ENV/bin/python"
     ./b2 || error "Manual boost compilation failed"
     ./b2 install || error "Manual boost installation failed"
     cd ..
